move utils to seperate file
